Fixed a warning in recent commit

+++ modules/juce_audio_basics/sources/juce_BufferingAudioSource.cpp
@@ -177,7 +177,7 @@ bool BufferingAudioSource::waitForNextAudioBlockReady (const AudioSourceChannelI
                 return true;
         }
 
-        if (! bufferReadyEvent.wait (endTime - now))
+        if (! bufferReadyEvent.wait (static_cast<int> (endTime - now)))
             return false;
 
         now = Time::getMillisecondCounter();
